pyproject.toml
[tool.poetry]
name = "Violet"
version = "0.0.0"
description = "A revolutionary website for all your nerdy friends, teams, organizations, clubs, or whatever!"
authors = ["Nate <minercraftcrusher100@gmail.com>"]

[tool.pylint.messages_control]
disable = ["too-few-public-methods"]

[tool.poetry.dependencies]
python = "^3.8"
fastapi = "0.73.0"
uvicorn = "^0.17.0"
PyJWT = "^2.3.0"
python-multipart = "^0.0.5"
requests = "^2.27.1"
postgrest-py = "^0.8.1"

[tool.poetry.dev-dependencies]
black = "^21.12b0"
isort = "^5.10.1"
mypy = "^0.931"
flake8 = "^4.0.1"
pylint = "^2.12.2"
pre-commit = "^2.17.0"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
